+++ source/fontcache.lua
@@ -1,11 +1,9 @@
-local gfx <const> = playdate.graphics
 local fonts = {}
 function getFont(t)
   if fonts[t] ~= nil then return fonts[t] end
-  if t == 'title' then
-    fonts[t] = gfx.font.new('font/Parodius Da!')
-  elseif t == 'ui' then
-    fonts[t] = gfx.font.new('font/whiteglove-stroked')
-  end
+  if t == 'menu' then
+    fonts[t] = playdate.graphics.font.new('font/Parodius Da!')
+  elseif t == 'textfield' then
+    fonts[t] = playdate.graphics.font.new('font/whiteglove-stroked')
   return fonts[t]
 end