+++ source/ui/menu.lua
@@ -1,65 +1,68 @@
-import 'CoreLibs/graphics'
-
 import 'fontcache'
 import 'textfield'
 
-local displayWidth, displayHeight = playdate.display.getSize()
-local centerX = displayWidth/2
-
 local gfx = playdate.graphics
 
-local menuOptions = {"START GAME", "OPTIONS"}
-local selection = 1
+Menu = {}
+Menu.__index = Menu
 
-local top = displayHeight/2-60
-local img = gfx.image.new(100, 22)
-gfx.pushContext(img)
-	gfx.fillRoundRect(0, 0, 100, 22, 4)
-gfx.popContext()
-local selected = gfx.sprite.new(img)
-selected:moveTo(centerX, top+selection*30)
-selected:add()
-local textFields = {}
-for i = 1, #menuOptions do
-	local tf = Textfield:new(centerX, top+i*30, menuOptions[i])
-	tf:setFont(getFont('menu'))
-	if i == selection then 
-		tf:setDrawMode(gfx.kDrawModeFillWhite)
-	else
-		tf:setDrawMode(gfx.kDrawModeCopy)
-	end
-	tf:add()
-	table.insert(textFields, tf)
-end
+function Menu:new(x, y, options, selection)
+	local self = {}
 
-function drawMenu()
-	selected:moveTo(centerX, top+selection*30)
-	for i = 1, #textFields do
-		local tf = textFields[i]
-		if i == selection then 
-			tf:setDrawMode(gfx.kDrawModeFillWhite)
-		else
-			tf:setDrawMode(gfx.kDrawModeCopy)
+	local offset = 30 -- due to 1-indexed arrays, first element already has offset applied
+	local buttonBg = nil
+	local textFields = {}
+
+	function self:add()
+		if buttonBg == nil then
+			local img = gfx.image.new(100, 22)
+			gfx.pushContext(img)
+				gfx.fillRoundRect(0, 0, 100, 22, 4)
+			gfx.popContext()
+			buttonBg = gfx.sprite.new(img)
+			buttonBg:moveTo(x, y+selection*offset)
+		end
+		buttonBg:add()
+		if #textFields == 0 then
+			for i = 1, #options do
+				local tf = Textfield:new(x, y+i*offset, options[i])
+				tf:setFont(getFont('menu'))
+				if i == selection then 
+					tf:setDrawMode(gfx.kDrawModeFillWhite)
+				else
+					tf:setDrawMode(gfx.kDrawModeCopy)
+				end
+				table.insert(textFields, tf)
+			end
+		end
+		for i = 1, #textFields do
+			textFields[i]:add()
 		end
 	end
-end
 
-function playdate.upButtonUp()
-	selection -= 1
-	if selection == 0 then 
-		selection = #menuOptions
+	function self:remove()
+		buttonBg:remove()
+		for i = 1, #textFields do
+			textFields[i]:remove()
+		end
 	end
-	drawMenu()
-end
 
-function playdate.downButtonUp()
-	selection += 1
-	if selection > #menuOptions then 
-		selection = 1
+	function draw()
+		buttonBg:moveTo(x, y+selection*offset)
+		for i = 1, #textFields do
+			local tf = textFields[i]
+			if i == selection then 
+				tf:setDrawMode(gfx.kDrawModeFillWhite)
+			else
+				tf:setDrawMode(gfx.kDrawModeCopy)
+			end
+		end
 	end
-	drawMenu()
-end
 
-function playdate.update()
-  playdate.graphics.sprite.update()
+	function self:setSelection(value)
+		selection = value
+		draw()
+	end
+
+	return self
 end